Add feature flag capability (#17)
* Add host_envs helper + specs

* Add FeatureSet and specs

* Add temp visual application check

* Update README regarding feature flags

Also changes some names to make them clearer

* Simplify HostEnv class

Renamed `ENV` to `ENV_NAME` as to be more explicit to what it is and also avoid possible name clashing.

Removed the `test?` method as for this (along with `development?`) is best to always use Rails.env, as HostEnv is about named environments and not if we are running in test/development or production mode.

* Remove config gem

Trying to have similar functionality without using a gem that seems to have quite a lot of dependencies.

* Refactor FeatureFlags class to remove Config gem dependency

Functionality remains almost the same (only thing remove is the `#disable!` method which I don't think is important) but has been refactored to not use the Config gem.

Co-authored-by: Jesus Laiz <zheileman@users.noreply.github.com>