CRIMAP-31 Add ping endpoint to expose build info
Ticket: https://dsdmoj.atlassian.net/browse/CRIMAP-31

This endpoint `/ping` exposes currently deployed build details and also can be used as a lightweight probe.

The ENV variables are exposed in the Dockerfile, and passed as `build-arg` when building the image in the deploy pipeline.

I've also pointed the k8s probes to the new /health endpoint created in PR #19.