CRIMAP-6 Configure Sentry exceptions reporting (#27)
As part of this I moved the 2 files in `lib/` to `app/lib` as that way Rails will load them automatically and we don't need to use an initialiser to require them.

I've created a team `laa-crime-apply` and added the app project to that team in Sentry with name `apply-for-criminal-legal-aid`.