Refinement to facilitate decorate collections (#79)
A simple refinement, that extends `Array` and `ActiveRecord::Relation` to facilitate the decoration of collection.

For example, given a decorator class named `CrimeApplicationDecorator` exists, we can decorate the list of applications in the dashboard by doing:

```ruby
@applications.decorate_each(CrimeApplicationDecorator) do |app|
```

The method yields a decorated instance. We can add to the decorator methods like `#applicant_name` (removing this method from the Applicant model as I don't think we use it anywhere else currently), a `#status_tag`, etc.

I will be doing the same with Presenters so maybe this instead of a decorator fits better the definition of presenter?

Refinement to facilitate present collections (#80)

Same as we did with decorators, but now also with presenters. It behaves and have the same interface.

Refer to PR #79 for more details.

Simplify collection refinements (#81)

It turned out we don't really need these to be pure refinements as we want these methods to be in all arrays and all AR relations, and we are accomplishing that just by including the modules.

Previous code was working because of that, not really because of the `refine`.

I would keep the structure as in the future we might have pure refinements that are to be used in specific classes or modules in a more surgical way.

w