CRIMRE 295 Ensure all applications are received (#230)
Receive the application if it has not yet been received.

In production, Review is notified of new applications via SNS.
As such, applications should be "received" before being returned by the
datastore.

This PR makes sure that receive_if_required! is called when loading from
the datastore to support other environments which do not have SNS set up,
and to act as a failsafe should an application be returned by the datastore
before its SNS event message has been processed.