CRIMAP-427 Bump schemas gem (appeal_lodged_date)
Bump the schemas gem, which contains the refactor of the appeal attributes, incorporating the new appeal_lodged_date.

NOTE: even if Review wasn't using these appeal attributes before, once Apply start submitting new applications without the old `appeal_with_changes_maat_id` attribute, if Review remains in an old version of the schemas gem, the structs will not be able to map these new applications, as that attribute was marked as "mandatory presence" and no longer is returned by the datastore.

This is the same issue we had recently with the `passportable` attribute.