Set default puma port to 3001 to avoid clashing (#388)
We run Apply on port 3000 and Datastore on port 3003 by default, either when running the rails application locally or when using docker-compose.

However Review was by default using port 3000 too, clashing with Apply when all 3 apps needed to be run to produce an end-to-end journey.

This change makes the port 3001 the default one for Review. It is still possible to locally override this by exposing `PORT=X` in your `.env.development.local` file if you are used to a different port.

Additionally support connections from _inside_ a docker container (for instance localstack) to the application running in the host machine (_not_ as a container).