Tweak docker-compose
With the objective of being able to run all 3 services (apply, review and datastore) at the same time as docker containers, some tweaks are needed.

Prometheus exporter port was hardcoded, it needs to be different in each container otherwise it will clash. Also by default now is disabled as most of the time we will never need it locally, but can be enabled for quick tests.

Added the datastore dummy credentials and endpoint (it works wether datastore is running also in a container, or on the host machine, assuming it's running in port 3003).

Added local docker to SNS bypass, for when Review is running as docker container.