Update deploy workflow with fix for ECR_URL
Same as we did in Datastore and Apply repos.

The previously used `secrets.ECR_URL` is now removed as part of ECR migration to short-term creds so is not available anymore to use in the ./.github/actions/deploy action.

However we can reference job outputs in other job steps to propagate this value.